Change file name for split functions better
