+++ lib/8583.js
@@ -2,7 +2,7 @@ const formats = require('./formats');
 const msgTypes = require('./msgTypes');
 const jxon = require('jxon');
 const types = require('./types');
-const requiredFields = require('./required');
+const requiredFields = require('./requiredFields');
 let {attachDiTimeStamps} = require('./helpers');
 let {validateFields, getLenType, getHex, getResType, getTransType, getAccType, getTranStatus} = require('./tools');
 const {validateSpecialFields, detectSpecial} = require('./specialFields/tools');
