+++ lib/requiredFields.js
@@ -11,6 +11,12 @@ module.exports = function(data, required_fields_file) {
 
   const missing_fields = filterMissingfields(data, required_fields);
 
+  if (missing_fields.length > 0) {
+    return {error: 'Processing code: ' + processing_code + ' - Missing required fields: ' + missing_fields};
+  }
+
+  return true;
+
   function findRequiredfields(arr, processing_code, message_code) {
     for (let i=0; i < arr.length; i++) {
       if (arr[i]['processing_code'] === processing_code) {
@@ -37,11 +43,5 @@ module.exports = function(data, required_fields_file) {
     });
     return missing_fields;
   }
-
-  if (missing_fields.length > 0) {
-    return {error: 'Processing code: ' + processing_code + ' - Missing required fields: ' + missing_fields};
-  } 
-
-  return true;
 
 };
